+++ Assets/Dust/Scripts/Editor/Actions/Core/DuInstantActionEditor.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-using UnityEditor;
-
-namespace DustEngine.DustEditor
-{
-    public abstract class DuInstantActionEditor : DuActionWithCallbacksEditor
-    {
-    }
-}